pr/requirements.txt
amqp>=5.2.0
aniso8601>=9.0.1
async-timeout>=4.0.3
bcrypt>=4.0.1
billiard>=4.2.0
blinker>=1.5
celery
chardet>=5.2.0
click>=8.1.3
click-didyoumean>=0.3.0
click-plugins>=1.1.1
click-repl>=0.3.0
dnspython>=2.3.0
email-validator>=1.3.1
Flask>=2.2.3
Flask-Caching>=1.10.0
Flask-CORS>=4.0.2
Flask-Excel>=0.0.7
Flask-Login>=0.6.2
Flask-Principal>=0.4.0
Flask-RESTful>=0.3.10
Flask-Security-Too>=5.1.2
Flask-Bcrypt>=1.0.1
Flask-SQLAlchemy>=3.0.3
Flask-WTF>=1.1.1
greenlet>=2.0.2
idna>=3.4
importlib-metadata>=6.0.0
itsdangerous>=2.1.2
Jinja2>=3.1.2
kombu>=5.3.3
lml>=0.1.0
matplotlib
MarkupSafe>=2.1.2
passlib>=1.7.4
weasyprint
prompt-toolkit>=3.0.40
pyexcel>=0.7.0
pyexcel-io>=0.6.6
pyexcel-webio>=0.1.4
python-dateutil>=2.8.2
pytz>=2023.3.post1
redis>=5.0.1
setuptools==75.6.0
six>=1.16.0
SQLAlchemy>=2.0.6
texttable>=1.7.0
typing-extensions>=4.5.0
tzdata>=2023.3
vine>=5.1.0
wcwidth>=0.2.9
Werkzeug>=2.2.3
WTForms>=3.0.1
zipp>=3.19.1
amqp==5.3.1
aniso8601==10.0.0
async-timeout==5.0.1
bcrypt==4.2.1
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
cachelib==0.9.0
celery==5.4.0
cffi==1.17.1
chardet==5.2.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.1
cssselect2==0.7.0
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-Excel==0.0.7
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.56.0
greenlet==3.1.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
kombu==5.4.2
lml==0.1.0
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.2
packaging==24.2
passlib==1.7.4
pillow==11.1.0
prompt_toolkit==3.0.50
pycparser==2.22
pydyf==0.11.0
pyexcel==0.7.1
pyexcel-io==0.6.7
pyexcel-webio==0.1.4
pyparsing==3.2.1
pyphen==0.17.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
setuptools==75.6.0
six==1.17.0
SQLAlchemy==2.0.38
texttable==1.7.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
weasyprint==64.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0
zopfli==0.2.3.post1
